erofs-utils: lib: support GNUTYPE_LONGNAME for tarerofs
The 'L' entry is present in a header for a series of 1 or more 512-byte
tar blocks that hold just the filename for a file or directory with a
name over 100 chars.

Following that series is another header block, in the traditional form:
   A header with type '0' (regular file) or '5' (directory), followed by
   the appropriate number of data blocks with the entry data.

In the header for this series, the name will be truncated to the 1st 100
characters of the actual name.

Tested-by: Jingbo Xu <jefflexu@linux.alibaba.com>
Signed-off-by: Gao Xiang <hsiangkao@linux.alibaba.com>
Link: https://lore.kernel.org/r/20230717073531.43203-1-hsiangkao@linux.alibaba.com